+++ backend/src/types.ts
@@ -19,6 +19,9 @@ const TYPES = {
   QuestionAiService: Symbol.for('QuestionAiService'),
   DuplicateService: Symbol.for('DuplicateService'),
   QueueService: Symbol.for('QueueService'),
+  RoleAssigneeService: Symbol.for('RoleAssigneeService'),
+  AllocationService: Symbol.for('AllocationService'),
+  ModeratorQueueService: Symbol.for('ModeratorQueueService'),
   AnswerService: Symbol.for('AnswerService'),
   ContextService: Symbol.for('ContextService'),
   CommentService: Symbol.for('CommentService'),
